HuangClaw/pyproject.toml
[project]
name = "huangclaw"
version = "0.1.0"
description = "A LangGraph coding agent with PDF RAG, Chroma storage, and Slack integration."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.0",
    "fastapi>=0.115.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.0",
    "pypdf>=5.0.0",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.10.0",
    "slack-bolt>=1.21.0",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
huangclaw = "huangclaw.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py311"
